requirements.txt
faiss-cpu==1.7.4
transformers==4.41.0
torch
sentence-transformers==2.6.1
